+++ lib/src/models/slash_command.dart
@@ -1,9 +1,10 @@
 import 'package:nyxx/nyxx.dart';
 // ignore: implementation_imports
 import 'package:nyxx/src/internal/cache/cacheable.dart';
+import 'package:nyxx_interactions/nyxx_interactions.dart';
+import 'package:nyxx_interactions/src/interactions.dart';
 import 'package:nyxx_interactions/src/models/command_option.dart';
-
-import 'package:nyxx_interactions/src/models/slash_command_type.dart';
+import 'package:nyxx_interactions/src/models/slash_command_permission.dart';
 
 abstract class ISlashCommand implements SnowflakeEntity {
   /// Unique id of the parent application
@@ -25,7 +26,23 @@ abstract class ISlashCommand implements SnowflakeEntity {
   Cacheable<Snowflake, IGuild>? get guild;
 
   /// Whether the command is enabled by default when the app is added to a guild
+  @Deprecated('Use canBeUsedInDm and requiresPermissions instead')
   bool get defaultPermissions;
+
+  /// Whether this slash command can be used in a DM channel with the bot.
+  bool get canBeUsedInDm;
+
+  /// A set of permissions required by users in guilds to execute this command.
+  ///
+  /// The integer to use for a permission can be obtained by using [PermissionsConstants]. If a member has any of the permissions combined with the bitwise OR
+  /// operator, they will be allowed to execute the command.
+  int get requiredPermissions;
+
+  /// If this command is a guild command, the permission overrides attached to this command, `null` otherwise.
+  Cacheable<Snowflake, ISlashCommandPermissionOverrides>? get permissionOverrides;
+
+  /// Get the permission overrides for this command in a specific guild.
+  Cacheable<Snowflake, ISlashCommandPermissionOverrides> getPermissionOverridesInGuild(Snowflake guildId);
 }
 
 /// Represents slash command that is returned from Discord API.
@@ -56,15 +73,39 @@ class SlashCommand extends SnowflakeEntity implements ISlashCommand {
 
   /// Whether the command is enabled by default when the app is added to a guild
   @override
+  @Deprecated('Use canBeUsedInDm and requiresPermissions instead')
   late final bool defaultPermissions;
 
-  /// Creates na instance of [SlashCommand]
-  SlashCommand(RawApiMap raw, INyxx client) : super(Snowflake(raw["id"])) {
+  /// Whether this slash command can be used in a DM channel with the bot.
+  @override
+  late final bool canBeUsedInDm;
+
+  /// A set of permissions required by users in guilds to execute this command.
+  ///
+  /// The integer to use for a permission can be obtained by using [PermissionsConstants]. If a member has any of the permissions combined with the bitwise OR
+  /// operator, they will be allowed to execute the command.
+  @override
+  late final int requiredPermissions;
+
+  @override
+  late final Cacheable<Snowflake, ISlashCommandPermissionOverrides>? permissionOverrides;
+
+  final Interactions _interactions;
+
+  /// Creates an instance of [SlashCommand]
+  SlashCommand(RawApiMap raw, this._interactions) : super(Snowflake(raw["id"])) {
     applicationId = Snowflake(raw["application_id"]);
     name = raw["name"] as String;
     description = raw["description"] as String;
     type = SlashCommandType(raw["type"] as int? ?? 1);
-    guild = raw["guild_id"] != null ? GuildCacheable(client, Snowflake(raw["guild_id"])) : null;
+    guild = raw["guild_id"] != null ? GuildCacheable(_interactions.client, Snowflake(raw["guild_id"])) : null;
+    canBeUsedInDm = raw["dm_permission"] as bool? ?? true;
+    requiredPermissions = int.parse(raw["default_member_permissions"] as String? ?? "0");
+
+    if (guild != null) {
+      permissionOverrides = SlashCommandPermissionOverridesCacheable(id, guild!.id, _interactions);
+    }
+
     defaultPermissions = raw["default_permission"] as bool? ?? true;
 
     options = [
@@ -72,4 +113,8 @@ class SlashCommand extends SnowflakeEntity implements ISlashCommand {
         for (final optionRaw in raw["options"]) CommandOption(optionRaw as RawApiMap)
     ];
   }
+
+  @override
+  Cacheable<Snowflake, ISlashCommandPermissionOverrides> getPermissionOverridesInGuild(Snowflake guildId) =>
+      SlashCommandPermissionOverridesCacheable(id, guildId, _interactions);
 }