+++ test/unit/event_controller.dart
@@ -1,4 +1,3 @@
-import 'package:nyxx_interactions/nyxx_interactions.dart';
 import 'package:nyxx_interactions/src/internal/event_controller.dart';
 import 'package:test/test.dart';